+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 - Game freeze when you start a fight without any player units
 - attackRange now depend on the chess tile size
 - tutorial is now the level 1
+- when an entity fails to spawn, we continue the spawn iteration instead of stopping it
 
 
 ## [0.1.4-alpha] - 2026-06-07
@@ -26,7 +27,7 @@ All notable changes to this project will be documented in this file.
 - log system with UI button
 - HP bar with different color for the player and enemy
 - new level & level is now handle with a json file
-- spawn algo for enemy, with differents trategy
+- spawn algo for enemy, with different strategy
 - use of json for the level generation
 - The level finish are in green